+++ lib/src/models/with_ancestor_ids.dart
@@ -1,3 +1,5 @@
+/// A tree node that can provide all its parents' IDs.
 abstract class WithAncestorIds<I> {
+  /// All parents' IDs.
   Iterable<I> get ancestorIds;
 }