+++ lib/src/models/with_children.dart
@@ -1,3 +1,5 @@
+/// Anything with children of a type [C].
 abstract class WithChildren<C> {
+  /// The children.
   List<C> get children;
 }