+++ lib/src/models/with_id_children_parent.dart
@@ -2,9 +2,6 @@ import 'with_children.dart';
 import 'with_id.dart';
 import 'with_parent.dart';
 
-abstract class WithIdChildrenParent<I, C, P> implements
-    WithId<I>,
-    WithChildren<C>,
-    WithParent<P>
-{
-}
+/// Anything with ID, children of type [C] and an optional parent of type [P].
+abstract class WithIdChildrenParent<I, C, P>
+    implements WithId<I>, WithChildren<C>, WithParent<P> {}