+++ lib/src/models/with_id_title.dart
@@ -1,6 +1,7 @@
 import 'with_id.dart';
 import 'with_title.dart';
 
+/// Anything with both an ID and a title.
 class WithIdTitle<T> implements WithId<T>, WithTitle {
   @override
   final T id;
@@ -8,6 +9,7 @@ class WithIdTitle<T> implements WithId<T>, WithTitle {
   @override
   final String title;
 
+  /// Creates an instance.
   const WithIdTitle(
     this.id,
     this.title,