+++ lib/src/models/with_id_title_children_parent.dart
@@ -1,8 +1,7 @@
-import 'with_id_title.dart';
 import 'with_id_children_parent.dart';
+import 'with_id_title.dart';
 
-abstract class WithIdTitleChildrenParent<I, C, P> implements
-    WithIdTitle<I>,
-    WithIdChildrenParent<I, C, P>
-{
-}
+/// Anything with ID, title, children of type [C],
+/// and an optional parent of type [P].
+abstract class WithIdTitleChildrenParent<I, C, P>
+    implements WithIdTitle<I>, WithIdChildrenParent<I, C, P> {}