+++ lib/src/models/with_id_title_homogenous_children_parent.dart
@@ -1,7 +1,10 @@
-import 'package:model_interfaces/model_interfaces.dart';
+import 'with_id_homogenous_children_parent.dart';
+import 'with_id_title_children_parent.dart';
 
-abstract class WithIdTitleHomogenousChildrenParent<I, T extends WithIdTitleHomogenousChildrenParent<I, T>> implements
-    WithIdTitleChildrenParent<I, T, T>,
-    WithIdHomogenousChildrenParent<I, T>
-{
-}
+/// Anything with ID, title, and children an an optional parent
+/// of the same type [T].
+abstract class WithIdTitleHomogenousChildrenParent<I,
+        T extends WithIdTitleHomogenousChildrenParent<I, T>>
+    implements
+        WithIdTitleChildrenParent<I, T, T>,
+        WithIdHomogenousChildrenParent<I, T> {}