+++ lib/src/models/with_parent.dart
@@ -1,3 +1,5 @@
+/// Anything with an optional parent of type [P].
 abstract class WithParent<P> {
+  /// The parent.
   P? get parent;
 }