+++ lib/src/models/with_title.dart
@@ -1,14 +1,17 @@
+/// Anything with a title.
 abstract class WithTitle {
+  /// The title.
   String get title;
 
+  /// Collects titles from an iterable.
   static Iterable<String> getTitles(Iterable<WithTitle> items) {
-    return items
-        .map((item) => item.title);
+    return items.map((item) => item.title);
   }
 
+  /// A case-insensitive comparator of titles.
   static int compareCaseInsensitive(WithTitle a, WithTitle b) {
     return a.title.toLowerCase().compareTo(
-      b.title.toLowerCase(),
-    );
+          b.title.toLowerCase(),
+        );
   }
 }